spark/requirements.txt
pyspark==3.2.0
delta-spark==1.2.1
pytest==8.0.0
chispa==0.10.0